# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the EMC Load Balancer Service.",
        "title": "EMC Load Balancer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/brands": {
            "get": {
                "description": "Get all brands",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brands"
                ],
                "summary": "List brands",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new brand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brands"
                ],
                "summary": "Create brand",
                "parameters": [
                    {
                        "description": "Brand details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.CreateBrandRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/brands/{id}": {
            "get": {
                "description": "Get a single brand by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brands"
                ],
                "summary": "Get brand by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-delete a brand",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brands"
                ],
                "summary": "Delete brand",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing brand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Brands"
                ],
                "summary": "Update brand",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Brand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.UpdateBrandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/coupons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all coupons (requires coupon:read permission)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "List coupons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new discount coupon (requires coupon:create permission)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "Create coupon",
                "parameters": [
                    {
                        "description": "Coupon details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.CreateCouponRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/coupons/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single coupon by its code (requires coupon:read permission)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "Get coupon by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coupon code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/coupons/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing coupon (requires coupon:update permission)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "Update coupon",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coupon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.UpdateCouponRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "emc_lb_src_pkg_entities.CreateBrandRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "banner": {
                    "type": "string",
                    "maxLength": 2048
                },
                "company_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "country": {
                    "type": "string",
                    "maxLength": 120
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "is_featured": {
                    "type": "boolean"
                },
                "logo": {
                    "type": "string",
                    "maxLength": 2048
                },
                "meta_description": {
                    "type": "string",
                    "maxLength": 500
                },
                "meta_keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meta_title": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "draft"
                    ]
                },
                "website": {
                    "type": "string",
                    "maxLength": 2048
                }
            }
        },
...
        "emc_lb_src_pkg_entities.CreateCouponRequest": {
            "type": "object",
            "required": [
                "code",
                "end_date",
                "start_date",
                "type",
                "value"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_discount_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "min_order_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "percentage",
                        "fixed_amount"
                    ]
                },
                "usage_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "emc_lb_src_pkg_entities.UpdateBrandRequest": {
            "type": "object",
            "properties": {
                "banner": {
                    "type": "string",
                    "maxLength": 2048
                },
                "company_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "country": {
                    "type": "string",
                    "maxLength": 120
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "is_featured": {
                    "type": "boolean"
                },
                "logo": {
                    "type": "string",
                    "maxLength": 2048
                },
                "meta_description": {
                    "type": "string",
                    "maxLength": 500
                },
                "meta_keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meta_title": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "draft"
                    ]
                },
                "website": {
                    "type": "string",
                    "maxLength": 2048
                }
            }
        },
...
        "emc_lb_src_pkg_entities.UpdateCouponRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_discount_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "min_order_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "percentage",
                        "fixed_amount"
                    ]
                },
                "usage_limit": {
                    "type": "integer",
                    "minimum": 0
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "emc_lb_src_pkg_res.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/emc_lb_src_pkg_res.ErrorCode"
                },
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/emc_lb_src_pkg_res.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {},
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "emc_lb_src_pkg_res.ErrorCode": {
            "type": "string",
            "enum": [
                "COMMON_BAD_REQUEST",
                "COMMON_NOT_FOUND",
                "COMMON_CONFLICT",
                "COMMON_INTERNAL_SERVER_ERROR",
                "COMMON_UNAUTHORIZED",
                "COMMON_FORBIDDEN",
                "COMMON_TOO_MANY_REQUESTS",
                "COMMON_INVALID_IDENTITY_TYPE"
            ],
            "x-enum-varnames": [
                "ErrCodeBadRequest",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeInternal",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeTooManyRequests",
                "ErrCodeInvalidIdentityType"
            ]
        },
        "emc_lb_src_pkg_res.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}